Refactor chronicle build flow into core stages
